chore!: change project base package name (#7)
Change the project base package name from `com.kori_47.utils` to
`io.github.kennedykori.utils`. This will satisfy
[this requirement](https://central.sonatype.org/publish/requirements/coordinates/#choose-your-coordinates)
necessary to publish the project on [Maven Central](https://central.sonatype.org/).

This change has been necessitated by the
[shutdown of Bintray](https://jfrog.com/blog/into-the-sunset-bintray-jcenter-gocenter-and-chartcenter/).
As such, the project will be migrating to Maven Central as the primary
hosting platform for the project binaries. Maven Central is a more
stable hosting service. Hopefully, this will result in a better and
more reliable experience.

BREAKING CHANGE: change base package name

A change of the base package name is a breaking change and should
result in a new major release.