build: set up publishing (#17)
Add publishing capabilities to the build system, allowing the library to
be consumed by other users/clients. The project can be published to the
following [Apache Maven](https://maven.apache.org/) repositories:

- A test repository relative to the build diretory(`build/repos/releases`).
- Maven local repository. See [here](https://maven.apache.org/repositories/local.html) for more details.
- [GitHub Packages](https://github.com/kennedykori/jutils/packages/2045860).
- [Maven Central](https://central.sonatype.com/artifact/io.github.kennedykori/utils/overview).

This is part of the project's efforts to migrate to Maven Central, as
highlighted in the issue [#13](https://github.com/kennedykori/jutils/issues/13).