ci: change prerelease versions suffix to rc (#36)
This is an attempt to improve the latest version selection on [Sonatype
OSS Nexus Repository](https://central.sonatype.org/publish/publish-guide/).
Currently, there is no guarantee that the latest version tags will be
updated on each deployment. See [this](https://support.sonatype.com/hc/en-us/articles/213464638-Why-are-the-latest-and-release-tags-in-maven-metadata-xml-not-being-updated-after-deploying-artifacts-)
for details.

This is an issue since some tools and services such as
[javadoc.io](https://javadoc.io/), seem to rely on these tags to select
the latest version of an artifact correctly. As such, this change tries
to address this by using prerelease versions that are friendly to Nexus
as described [here](https://sonatype.github.io/sonatype-aether/apidocs/org/sonatype/aether/util/version/GenericVersionScheme.html).
Specifically, this changes from using `dev` as a prerelease tag to `rc`.

Hopefully, with this change, if the `maven-metadata.xml` gets rebuilt
by Nexus, the latest tag will be updated correctly. This
[issue](https://github.com/ctapobep/blog/issues/3) provides more
details on how and who updates the `maven-metadata.xml` file.